requirements.txt
streamlit==1.27.2
pandas
xgboost
joblib
python-dotenv
